Add Data infrastructure.

+++ src/Catering.Application/FilterResult.cs
@@ -5,5 +5,5 @@ public class FilterResult<T>
     public IEnumerable<T> Result { get; set; }
     public short PageSize { get; set; }
     public short PageIndex { get; set; }
-    public short TotalNumberOfPages { get; set; }
+    public int TotalNumberOfPages { get; set; }
 }
